Convert String.toInt exceptions into Options
Wrapping toInt with Option does not prevent exceptions.
It is clear in the slides that we are combining Monads
but this use of Option might confuse some audience if they
are new to Monads but versed on Try/Option.

Unfortunately, to make this proper I cannot think of a less intrusive
option than wrapping toInt on Try and getting the Option.

My proposed change muddles the clarity of the code on the slide but it has
the proper behaviour so the message will be conveyed
if somebody tries ```run("a")```.